Rename slackbot dir
