+++ lovebot/DumbBot.js
