+++ lovebot/LoveBot.js
