+++ lovebot/SlackBot.js
