+++ lovebot/index.js
@@ -1,4 +1,4 @@
-const LoveBot = require('./slackbot/LoveBot');
+const LoveBot = require('./LoveBot');
 
 
 const bot = new LoveBot(process.env.SLACK_TOKEN);
