+++ lovebot/responses.js
